backend/requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-groq
langchain-community

faiss-cpu

openai
tiktoken